+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to `Combustor` will be documented in this file.
 - Command for creating `combustor.yml` configuration file
 - Fields that can be excluded through `excluded_files`
 - Specify customized `application` path using `app_path` 
+- `create:repository` for creating Entity Repositories
 
 ### Changed
 - Code coverage provider to `Codecov`